ConchingModel/requirements.txt
astroid==2.9.0
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
cycler==0.11.0
dill==0.3.4
et-xmlfile==1.1.0
fonttools==4.28.2
iniconfig==1.1.1
isort==5.10.1
joblib==1.1.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
matplotlib==3.5.0
mccabe==0.6.1
multiprocess==0.70.12.2
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pathos==0.2.8
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
pox==0.3.0
ppft==1.6.6.4
py==1.11.0
pyaml==21.10.1
pylint==2.12.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
SALib==1.4.5
scikit-learn==1.0.1
scikit-optimize==0.9.0
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
snakeviz==2.1.1
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
wrapt==1.13.3
